requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
ecdsa==0.19.2
fastapi==0.136.1
h11==0.16.0
idna==3.13
passlib==1.7.4
psycopg2-binary==2.9.12
pyasn1==0.6.3
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-jose==3.5.0
redis==7.4.0
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
